Crate graphql_ws_client

graphql-ws-client

graphql-ws-client implements asynchronous GraphQL-over-Websocket using the graphql-transport-ws protocol. It tries to be websocket client, graphql client and async executor agnostic and provides built in support for:

Cynic & Graphql-Client GraphQL clients.

async-tungstenite & ws-stream-wasm Websocket Clients .

The async-std & tokio async runtimes.

If you’d like to use another client or runtime adding support should hopefully be trivial.

Modules

graphql

This module contains traits that abstract over GraphQL implementations, allowing this library to be used with different GraphQL client libraries.

websockets

Contains traits to provide support for various underlying websocket clients.

Structs

AsyncWebsocketClient

A websocket client

AsyncWebsocketClientBuilder

A websocket client builder

SubscriptionStream

A futures::Stream for a subscription.

Enums

Error

Error type

Type Aliases

GraphQLClientClient

A websocket client for the graphql_client graphql crate

GraphQLClientClientBuilder

A websocket client builder for the graphql_client graphql crate