graphql_ws_client
Module graphql

This module contains traits that abstract over GraphQL implementations, allowing this library to be used with different GraphQL client libraries.

Support is provided for cynic & graphql_client, but other client libraries can be added by implementing these traits for those libraries.

Structs

GraphQLClient

Provides an implementation of GraphqlClient for the graphql_client GraphQL crate

StreamingOperation

A streaming operation for a GraphQLQuery

Traits

GraphqlClient

A trait for GraphQL clients.

GraphqlOperation

An abstraction over GraphQL operations.