Trait graphql_ws_client::websockets::WebsocketMessage

pub trait WebsocketMessage: Debug {
    type Error: Error + Send + 'static;

    // Required methods
    fn new(text: String) -> Self;
    fn text(&self) -> Option<&str>;
    fn error_message(&self) -> Option<String>;
    fn is_ping(&self) -> bool;
    fn is_pong(&self) -> bool;
    fn is_close(&self) -> bool;
}

An abstraction around WebsocketMessages

graphql-ws-client doesn’t implement the websocket protocol itself. This trait provides part of the integration with websocket client libraries.

Required Associated Types

type Error: Error + Send + 'static

The Error type for this websocket client.

Required Methods

fn new(text: String) -> Self

Constructs a new message with the given text

fn text(&self) -> Option<&str>

Returns the text (if there is any) contained in this message

fn error_message(&self) -> Option<String>

Returns the text (if there is any) of the error

fn is_ping(&self) -> bool

Returns true if this message is a websocket ping.

fn is_pong(&self) -> bool

Returns true if this message is a websocket pong.

fn is_close(&self) -> bool

Returns true if this message is a websocket close.

Object Safety

This trait is not object safe.