Trait graphql_ws_client::graphql::GraphqlClient

pub trait GraphqlClient {
    type Response: DeserializeOwned + Send;
    type DecodeError: Error + Send + 'static;

    // Required method
    fn error_response(
        errors: Vec<Value>,
    ) -> Result<Self::Response, Self::DecodeError>;
}

A trait for GraphQL clients.

Required Associated Types

type Response: DeserializeOwned + Send

The generic response type for this GraphqlClient implementation

Our client will decode this, then pass it to a GraphqlOperation for decoding to the specific response type of the GraphqlOperation.

type DecodeError: Error + Send + 'static

The error that will be returned from failed attempts to decode a Response.

Required Methods

fn error_response( errors: Vec<Value>, ) -> Result<Self::Response, Self::DecodeError>

Decodes some error JSON into a Response

Object Safety

This trait is not object safe.